requirements.txt
black==21.11b1
django<4.1
factory-boy==3.2.1
flake8-bugbear==21.9.2
flake8-django==1.1.1
flake8==4.0.1
isort==5.10.1
nox==2021.10.1
PyLTI1p3==1.9.1
pytest-django==4.5.1
pytest==6.2.5
